First commit. Not working yet.

+++ src/ukf.h
@@ -65,7 +65,7 @@ public:
   int n_aug_;
 
   ///* Sigma point spreading parameter
-  double lambda_;
+  int lambda_;
 
 
   /**
